requirements/development.txt
astroid==2.3.2
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.8
flake8-import-order==0.18.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.3
pytz==2019.3
requests==2.22.0
rope==0.14.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
uWSGI==2.0.18
wrapt==1.11.2
